requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
openai
httpx
python-dotenv
psycopg2-binary
sqlalchemy
uvicorn[standard]
python-jose
PyYAML
typing-extensions
dpath
bcrypt
fastapi==0.110.0
uvicorn==0.29.0
asyncpg==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.0
sqlalchemy==2.0.30
alembic==1.13.1
pyyaml==6.0.1
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.28.0
psycopg2-binary>=2.9.9
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
openai>=1.3.0
httpx>=0.25.1
python-multipart>=0.0.6
